feat(gpu/qrm): add duration metrics for allocation stages
Add three new metrics to track filter, sort and bind durations during GPU allocation. Wrap the corresponding allocation strategy steps with timing wrappers that emit metrics using the existing metrics emitter, and add comprehensive unit tests for the timing utilities.

+++ pkg/agent/qrm-plugins/gpu/staticpolicy/policy.go
@@ -727,7 +727,7 @@ func (p *StaticPolicy) AllocateAssociatedDevice(
 
 		elapsed := time.Since(startTime)
 		_ = p.emitter.StoreFloat64(util.MetricNameAllocateAssociatedDeviceDuration,
-			float64(elapsed/time.Millisecond), metrics.MetricTypeNameRaw,
+			float64(elapsed)/float64(time.Millisecond), metrics.MetricTypeNameRaw,
 			metrics.MetricTag{Key: "deviceName", Val: req.DeviceName},
 			metrics.MetricTag{Key: "accompanyResourceName", Val: req.AccompanyResourceName},
 			metrics.MetricTag{Key: "success", Val: strconv.FormatBool(respErr == nil)},
